moved multinest specific functions to multinest
